+++ src/pages/UnitsPage/UnitsListColumns.jsx
@@ -1,7 +1,11 @@
 import React from "react";
 import { IconButton } from "@mui/material";
 import DeleteRoundedIcon from "@mui/icons-material/DeleteRounded";
+import { useGetProfileQuery } from "../../redux/auth/authApiSlice";
+
 export const ComposeColumns = (onDelete) => {
+  const { data, isSuccess } = useGetProfileQuery();
+
   return [
     {
       Header: "Name",
@@ -63,16 +67,23 @@ export const ComposeColumns = (onDelete) => {
       id: "action",
       width: 128,
       sortable: false,
-      Cell: ({ row }) => (
-        <div className="flex items-center">
-          <IconButton
-            className="p-4"
-            onClick={(ev) => onDelete(ev, row.original)}
-          >
-            <DeleteRoundedIcon color="action" />
-          </IconButton>
-        </div>
-      ),
+      Cell: ({ row }) => {
+        if (isSuccess)
+          if (
+            data?.Role?.Role_Resources?.find((x) => x?.resource?.Name == "Unit")
+              ?.Delete == true
+          )
+            return (
+              <div className="flex items-center">
+                <IconButton
+                  className="p-4"
+                  onClick={(ev) => onDelete(ev, row.original)}
+                >
+                  <DeleteRoundedIcon color="action" />
+                </IconButton>
+              </div>
+            );
+      },
     },
   ];
 };
